Resource api class and method changes
